refactor income popup handling to separate file
